AI healthcare Diagnosis final/AI healthcare Diagnosis (2)/AI healthcare Diagnosis/requirements.txt
flask==2.0.1
tensorflow==2.15.0
numpy>=1.19.5
Pillow>=8.4.0
opencv-python>=4.5.4.60
python-dotenv>=0.19.2
pytesseract>=0.3.8
pdf2image>=1.16.0
werkzeug==2.0.2 